Some rebase fixing and restructuring
